+++ lib/lex_llm/chunk.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module LexLLM
-  class Chunk < Message
-  end
-end